backend/requirements.txt
gspread
google-generativeai
pandas
oauth2client
python-dotenv
requests
beautifulsoup4
lxml
